apps/telegram_bot/bot.py
"""
Telegram bot implementation.
"""

from django.conf import settings
from telegram import Update
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes
import logging

logger = logging.getLogger(__name__)


async def start_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle /start command."""
    await update.message.reply_text(
        "Hello! I'm the Noti bot. I'll help you manage your notifications."
    )


async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle /help command."""
    help_text = """
Available commands:
/start - Start the bot
/help - Show this help message
/notifications - Show your notifications
/settings - Manage your settings
    """
    await update.message.reply_text(help_text)


async def handle_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle regular messages."""
    message = update.message.text
    await update.message.reply_text(f"You said: {message}")


def handle_telegram_update(update_data):
    """Handle incoming Telegram update."""
    # This is a placeholder for now
    # In a real implementation, you would process the update here
    logger.info(f"Received Telegram update: {update_data}")


# Bot setup (this would be called when the bot starts)
def setup_bot():
    """Setup the Telegram bot."""
    if not settings.TELEGRAM_BOT_TOKEN:
        logger.warning("TELEGRAM_BOT_TOKEN not set, bot will not start")
        return None
    
    application = Application.builder().token(settings.TELEGRAM_BOT_TOKEN).build()
    
    # Add handlers
    application.add_handler(CommandHandler("start", start_command))
    application.add_handler(CommandHandler("help", help_command))
    application.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, handle_message))
    
    return application
